refactor: break down WhitelistContractsCard
